Break ElizaEmailNotifierActor into an abstract class and two subclasses
